Do not generate puzzles with too many singular/plural pairs
Previously, some of the generated puzzles were tiresome due to too
many pairs of words which were identical except an added -S or -ES.

With this change, puzzles which have more than MAX_PLURALS pairs will
be rejected. (See params.py for definition of MAX_PLURALS). To disable
completely, the CAP_PLURALS parameter may be set to False.

Note that the word "plural" is used loosely here as we only care about
removing the boring cases. In particular, when the suffixes -S or -ES
make a noun trivially plural (or a verb trivially singular).

Here's a typical example of a puzzle with too many pairs:

	['APPLAUSE', 'GAUGE', 'GAUGES', 'GLUE', 'GLUES', 'GUESS', 'GUESSES', 'GULL', 'GULLS', 'GULP', 'GULPS', 'LEAGUE', 'LEAGUES', 'LUGGAGE', 'LUGS', 'LULL', 'LULLS', 'PAUSE', 'PAUSES', 'PLAGUE', 'PLAGUES', 'PLUG', 'PLUGS', 'PLUS', 'PLUSES', 'PULL', 'PULLS', 'PULP', 'PULPS', 'PULSE', 'PULSES', 'PUPS', 'PUSS', 'PUSSES', 'SAUSAGE', 'SAUSAGES', 'SLUG', 'SLUGS', 'SUES', 'SUPPLE', 'USAGE', 'USAGES', 'USELESS', 'USES', 'USUAL']

About the algorithm:

We iterate over each word that ends in -S or -ES, remove the suffix,
and then check if the base word is also in the list. Despite being
overly simplistic, this seems to work well in practice.

However, there do exist some words which the algorithm misidentifies
as singular/plural pairs:

	BRAS/BRASS
	DISCUS/DISCUSS

There are others which are technically wrong, but may not be an issue:

	ADVENTURES/ADVENTURESS
	ASSES/ASSESS
	CARES/CARESS
	DEADLINES/DEADLINESS
	LARGES/LARGESS
	NEEDLES/NEEDLESS
	OGRES/OGRESS
	POSSES/POSSESS
	PRINCES/PRINCESS
	SALTINES/SALTINESS
	TIMELINES/TIMELINESS

For this list, the count of pairs would probably exceed the cap even
without an incorrect +1 from the misidentification. For example, just
from the letters A, E, and S, one already has a count of 6: ASS/ASSES,
EASE/EASES, SASS/SASSES, SEA/SEAS, SEE/SEES, ASSESS/ASSESSES.