Remove puzzles that have too many singular/plural pairs
The vast majority of the puzzles with an 'S' end up with a too many
words being trivially the same: one is singular and the other plural.
Before pruning down to a max of 3 such pairs, over 1/3rd of the games
(1108/3000) had an 'S' in them. Afterward, only 1/300th of the
generated puzzles (11/3000) contained an 'S'.