Implement PRINT_INVALID="why" to show % reason why invalid
Replaces PRINT_INVALID="progress", which was not as helpful because it
showed only the first reason a set of letters was rejected. Now, we're
adding fractional amounts if multiple reasons are true.

Note that this does technically make the algorithm slower as it
continues checking for various ways the set of letters is invalid.
However, there are much bigger bottlenecks to the speed of this
generation algorithm and has not been a serious issue in practice.