Games regenerated with reduced gerund pairs
Now no games exist in which more than five pairs of verbs / gerundive
nouns exist. (E.g., JUMP / JUMPING). This improves the quality of play
significantly.

However, that restriction reduced the number of available games to a
little over 2000. Since I would like there to be at least 3000 games
available, I tweaked params.py and widened the allowable number of
words (min 25, from 30) and points (min 60, was 80). This has had no
negative side-effects in terms of playability or enjoyment. Possibly
this is because the scoring system is rather simplistic and not
strongly correlated with difficulty. (Maybe score should not even be
considered for limiting games?)