Implement ./utils.py slook <word>
Search for a prefix in the SCOWL dictionaries and return results
sorted by most to least common.

Usage: ./utils.py slook <pattern>

Shows words matching the given prefix (regular expression ^pattern.*)
and which wordlist file it was found in. The number at the end of the
filename is an indication of how common SCOWL thinks the word is. This
is useful to see why a word is getting rejected. The default threshold
is <=35, so english-words.35 is used but english-words.40 is not.

    $ ./utils.py slook alfalfa
    word_lists/scowl/english-words.40: alfalfa
    word_lists/scowl/english-words.80: alfalfas

Too many matches? For the exact word, put a '$' at the end. For example,

    $ ./utils.py slook mes | wc -l
    914
    $ ./utils.py slook mes$
    word_lists/scowl/english-words.35: mes