make_puzzles(): preterite count, manual letters, print counts
PRETERITE: Simple Past Tense

Too many pairs of words of the form "YODEL"/"YODELED" are now marked
invalid. Like -S and -ING, lots of -ED is not terribly fun. (Try
playing DEFLOUY).

____

AUTO/MANUAL LETTER SELECTION

When letters are specified, `manual_puzzle` is usually set to True.
That was a problem when regenerating existing puzzle files as they
would appear to be manually generated.

The previous workaround of --force-auto and the force_auto variable
was clunky and hard to understand.

Now, make_puzzles() simply preserves the manual_puzzles variable from
existing puzzle files.

____

When in CSV output mode, counts of pairs that might rule out the
puzzle are now printed.