Save some of the calculated quality metrics in .json file
The theory being we can use that information later at runtime to
quickly filter through games based on criteria.

Note that the values were added as subfields to generation_info. Since
it is a "dict" type in the Puzzle object, this is not a breaking
change to the .json format: old files can be loaded just fine and old
versions of the program should be able to handle the newer format as
well.

    quality = {
        'is_valid'           : is_valid,
        's_pairs'            : s_pairs,
        'ed_pairs'           : ed_pairs,
        'ing_pairs'          : ing_pairs,
        'uniqueness'         : utils.uniqueness(results),
    }