Restored handpicked selection of simple past tense puzzles
Previously, we removed over a thousand puzzles which had more than
seven pairs of a verb and it preterite form (verb+ed).

This adds about a dozen puzzles back in even though they have 8 or 9
such pairs. All of them have pangrams which are unique and have a
fairly good mix of interesting words.